simple-data-analytic/backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pandas==2.2.3
numpy==1.26.4
openpyxl==3.1.2
python-multipart==0.0.9
openai==1.35.0
scikit-learn==1.5.0
matplotlib==3.9.0
seaborn==0.13.2
plotly==5.22.0
python-dotenv==1.0.1
